pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "malleus-dev"
version = "0.1.0"
description = "Root ontology in LinkML for ontology-typed knowledge graphs with write-time validation and content-addressable convergence"
readme = "README.md"
license = { text = "Apache-2.0" }
authors = [{ name = "Luis Guzman Lorenzo" }]
requires-python = ">=3.10"
keywords = ["ontology", "knowledge-graph", "linkml", "typed-graph", "crdt", "schema"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "linkml>=1.10",
    "linkml-runtime>=1.10",
    "networkx>=3.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
prolog = ["pyswip>=0.3"]
dev = ["pytest>=8", "pytest-cov>=5", "build>=1.0", "twine>=5.0"]

[project.urls]
Homepage = "https://github.com/kieleth/malleus-dev"
Documentation = "https://github.com/kieleth/malleus-dev/blob/main/docs/ARCHITECTURE.md"
Repository = "https://github.com/kieleth/malleus-dev"
Issues = "https://github.com/kieleth/malleus-dev/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/malleus"]

[tool.hatch.build.targets.wheel.shared-data]
"ontology" = "share/malleus/ontology"

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/ontology",
    "/tests",
    "/docs",
    "/README.md",
    "/LICENSE",
    "/CHANGELOG.md",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
